Break up Sensei plan step
